Refactor `MultiSig` / `Tx` defs to `Tx` module
- move transaction definitions from `UTxO` to own module
- add hashScript to `MultiSignatureScript` typeclass
- define first tag for native multi-signature
- move address defintions to `Address module`
