+++ shelley/chain-and-ledger/executable-spec/delegation.cabal
@@ -17,7 +17,8 @@ flag development
     manual: True
 
 library
-  exposed-modules:   BaseTypes
+  exposed-modules:   Address
+                     BaseTypes
                      BlockChain
                      Coin
                      Keys
@@ -26,10 +27,10 @@ library
                      PParams
                      EpochBoundary
                      LedgerState
-                     MultiSig
                      Delegation.PoolParams
                      Delegation.Certificates
                      OCert
+                     Tx
                      Updates
                      STS.Avup
                      STS.Bbody
