+++ shelley/chain-and-ledger/executable-spec/src/EpochBoundary.hs
@@ -30,6 +30,7 @@ module EpochBoundary
   , groupByPool
   ) where
 
+import           Address
 import           Coin
 import           Delegation.Certificates (StakeKeys (..), StakePools (..),
                                           decayKey, decayPool, refund)
@@ -37,6 +38,7 @@ import           Delegation.PoolParams   (RewardAcnt (..), PoolParams(..))
 import           Keys
 import           PParams hiding (a0, nOpt)
 import           Slot
+import           Tx
 import           UTxO hiding (dom)
 
 import qualified Data.Map.Strict         as Map
