+++ shelley/chain-and-ledger/executable-spec/src/STS/Ledgers.hs
@@ -16,7 +16,7 @@ import           Keys
 import           LedgerState
 import           PParams
 import           Slot
-import           UTxO
+import           Tx
 
 import           Control.State.Transition
