+++ shelley/chain-and-ledger/executable-spec/src/STS/Utxo.hs
@@ -23,6 +23,7 @@ import           Keys
 import           LedgerState hiding (dms)
 import           PParams
 import           Slot
+import           Tx
 import           Updates
 import           UTxO
