+++ shelley/chain-and-ledger/executable-spec/test/MockTypes.hs
@@ -4,6 +4,7 @@ import           Cardano.Crypto.DSIGN (MockDSIGN)
 import           Cardano.Crypto.Hash (ShortHash)
 import           Cardano.Crypto.KES (MockKES)
 
+import qualified Address
 import qualified BlockChain
 import qualified Delegation.Certificates
 import qualified Delegation.PoolParams
@@ -11,6 +12,7 @@ import qualified Keys
 import qualified LedgerState
 import qualified OCert
 import qualified STS.Chain
+import qualified Tx
 import qualified UTxO
 
 type DCert = Delegation.Certificates.DCert ShortHash MockDSIGN
@@ -37,15 +39,15 @@ type UTxOState = LedgerState.UTxOState ShortHash MockDSIGN
 
 type DPState = LedgerState.DPState ShortHash MockDSIGN
 
-type Addr = UTxO.Addr ShortHash MockDSIGN
+type Addr = Address.Addr ShortHash MockDSIGN
 
-type Tx = UTxO.Tx ShortHash MockDSIGN
+type Tx = Tx.Tx ShortHash MockDSIGN
 
-type TxBody = UTxO.TxBody ShortHash MockDSIGN
+type TxBody = Tx.TxBody ShortHash MockDSIGN
 
-type TxIn = UTxO.TxIn ShortHash MockDSIGN
+type TxIn = Tx.TxIn ShortHash MockDSIGN
 
-type TxOut = UTxO.TxOut ShortHash MockDSIGN
+type TxOut = Tx.TxOut ShortHash MockDSIGN
 
 type UTxO = UTxO.UTxO ShortHash MockDSIGN
